fix tests and refactor
