+++ src/main/java/uk/gov/hmcts/reform/iacasepaymentsapi/domain/handlers/presubmit/PaymentAppealHandler.java
@@ -116,10 +116,7 @@ public class PaymentAppealHandler implements PreSubmitCallbackHandler<AsylumCase
             );
         asylumCase.write(FEE_PAYMENT_APPEAL_TYPE, YesOrNo.NO);
 
-        Fee feeSelected = null;
-        if (!FeesHelper.feeExistsForDecisionType(asylumCase)) {
-            feeSelected = FeesHelper.findFeeByHearingType(feeService, asylumCase);
-        }
+        Fee feeSelected = FeesHelper.findFeeByHearingType(feeService, asylumCase);
 
         log.info("Selected Fee for caseId: {}, {}", caseId, feeSelected);
